feat: Rough working version with prop drilling
